Upgrade on the HttpMethod decorator

+++ backend/src/controller/Orders.controller.ts
@@ -1,53 +1,54 @@
-import { Request, Response } from "express-serve-static-core";
-import { Controller } from "./Controller";
-import { OrdersQueue, Order } from "../model/Order";
+import { Request, Response } from "express";
+import { OrderQueue, Order } from "../model/Order";
 import { Customer } from "../model/Customer";
 import uuid = require("uuid/v4");
 import { EmailAddress, PhoneNum } from "../type/Location";
 import { Money } from "../type/Money";
 import { FieldDef, QueryResult } from "pg";
-import { Paypal } from "../model/Paypal";
+import { PaypalSetup, Paypal } from "../model/Paypal";
 import { Shipping } from "../model/Shipping";
 import { wsServer } from "../index";
 import { OrderConstructor, ExistingOrderBody } from "../interface/Order.interface";
+import { OrderEvent } from "../event/Order.event";
 
-export class OrdersController extends Controller
+export class OrdersController
 {
-  private static unprocessed: OrdersQueue = new OrdersQueue([]);
-  public static async add(req: Request, res: Response): Promise<void>
+  private static _orderLeft: OrderQueue = new OrderQueue();
+  private static _events: OrderEvent = new OrderEvent();
+  public static add(req: Request, res: Response): void
   {
     try {
-      const order: Order = Order.generate(req.body, req);
-      this.unprocessed.enqueue(order);
-      ws.emit(, order.toPrimObj());
+      const order: Order = Order.generate(req.body.order, req);
+      this._orderLeft.enqueue(order);
+      this._events.add(order);
+      order.save();
+      res.send({ success: true });
     } catch (e) {
-      const ex: Error = e;
-      hconsole.error(ex);
-      res.send({success: false, error: "Unable to add order into the system."});
+      res.sendError(e, "System was unable to add the order.");
     }
   }
-  public static async complete(req: Request, res: Response): Promise<void>
+  public static complete(req: Request, res: Response): void
   {
     try {
-      if(req.body.orderID !== this.unprocessed.getNextOrderID())
-        throw new Error("Order ID wasn't next in the queue.");
-      const order = this.unprocessed.dequeue();
-      const shipping = order.getShipping();
+      if(req.body.order.id !== this._orderLeft.getNextOrderID())
+        throw new Error("System cannot get an order that isn't next in the queue.");
+      const order: Order = this._orderLeft.dequeue();
+      const shipping: Shipping = order.getShipping();
+      order.complete();
+      shipping.complete();
       order.save();
       shipping.save();
       res.send({success: true});
     } catch (e) {
-      const ex: Error = e;
-      hconsole.error(ex);
-      res.send({success: false, error: "Unable to complete the next order!"});
+      res.sendError(e, "System was unable to complete the next order!");
     }
   }
   public static hold(req: Request, res: Response): void
   {
     try {
-      if(req.body.orderID === this.unprocessed.getNextOrderID())
+      if(req.body.orderID === this._orderLeft.getNextOrderID())
         throw new Error("The order id isn't next in the queue");
-      this.unprocessed.hold();
+      this._orderLeft.hold();
       res.send({success: true});
     } catch (e) {
       const ex: Error = e;
@@ -71,13 +72,24 @@ export class OrdersController extends Controller
   public static remove(req: Request, res: Response): void
   {
     try {
-      const order: Order = Order.From.id(req.body.id);
+      if (!req.body.order.id)
+        throw new Error("Client didn't provide an order id to remove.");
+      const order: Order = Order.From.id(req.body.order.id);
+      const orderID: string = order.getID();
+      const orders: Order[] = this._orderLeft.getAllOrders();
+      for (let i = 0; i < orders.length; i++) {
+        const cur: Order = orders[i];
+        if (cur.getID() === orderID) {
+          orders.splice(i, 1);
+          break;
+        }
+        if (i === orders.length)
+          throw new Error("System couldn't find the order the unprocessed order.");
+      }
       order.delete();
       res.send({success: true});
     } catch (e) {
-      const ex: Error = e;
-      hconsole.error(ex);
-      res.send({success: false, error: "System was unable to delete the order."});
+      res.sendError(e, "System was unable to remove the order.");
     }
   }
   public static update(req: Request, res: Response): void
@@ -116,18 +128,10 @@ export class OrdersController extends Controller
   }
   public static feed(req: Request, res: Response)
   {
-    ws.on("connection", wss =>
-    {
-      wss.on("message", message =>
-      {
-        try {
-          const msg = message;
-          wss.send(msg);
-        } catch (e) {
-          const ex: Error = e;
-          hconsole.error(ex);
-        }
+    wsServer.once("connection", client => {
+      this._events.on(this._events.getAdd(), (order: Order) => {
+        client.send(order.toPrimObj());
       });
-    })
+    });
   }
 }
